Unify project with flexible-rating and m.e.i.n.e.l in mind
* Migrate tests to use wct-mocha
* Rework localization to be a mixin
* Add demo for all elements
* Cleanup code
* Ensure compatibility with other browsers such as IE
