+++ .eslintrc.json
@@ -8,31 +8,73 @@
     "no-console": "off",
     "array-callback-return": "warn",
     "indent": "off",
-    "indent-legacy": ["error", 4, { "SwitchCase": 1 }],
+    "indent-legacy": [
+      "error",
+      4,
+      {
+        "SwitchCase": 1
+      }
+    ],
     "no-multi-spaces": "error",
-    "linebreak-style": ["error", "unix"],
-    "semi": ["error", "always"],
+    "linebreak-style": [
+      "error",
+      "unix"
+    ],
+    "semi": [
+      "error",
+      "always"
+    ],
     "no-sequences": "error",
     "comma-spacing": "error",
-    "comma-dangle": ["error", "never"],
+    "comma-dangle": [
+      "error",
+      "never"
+    ],
     "dot-notation": "error",
-    "quotes": ["error", "single"],
-    "brace-style": ["error", "1tbs", { "allowSingleLine": true }],
+    "quotes": [
+      "error",
+      "single"
+    ],
+    "brace-style": [
+      "error",
+      "1tbs",
+      {
+        "allowSingleLine": true
+      }
+    ],
     "camelcase": "error",
     "no-var": "error",
-    "no-unused-vars": ["error", {"vars": "all", "varsIgnorePattern": "MammoocLocalizeBehavior"}],
     "block-scoped-var": "error",
-    "consistent-this": ["warn", "that"],
+    "consistent-this": [
+      "warn",
+      "that"
+    ],
     "eqeqeq": "error",
     "guard-for-in": "warn",
     "object-shorthand": "error",
     "arrow-spacing": "error",
     "prefer-arrow-callback": "error",
-    "arrow-body-style": ["error", "as-needed"],
+    "arrow-body-style": [
+      "error",
+      "as-needed"
+    ],
     "valid-jsdoc": "error",
-    "capitalized-comments": ["warn", "always", {"ignoreConsecutiveComments": true}],
+    "capitalized-comments": [
+      "warn",
+      "always",
+      {
+        "ignoreConsecutiveComments": true
+      }
+    ],
     "no-global-assign": "warn",
-    "strict": ["warn", "global"]
+    "strict": [
+      "warn",
+      "global"
+    ],
+    "object-curly-spacing": [
+      "error",
+      "always"
+    ]
   },
   "settings": {
     "html/indent": "+4",
@@ -45,8 +87,7 @@
     "Polymer": false,
     "define": false,
     "require": false,
-    "exports": false,
-    "MammoocLocalizeBehavior": true
+    "exports": false
   },
   "parserOptions": {
     "sourceType": "module"
