+++ .travis.yml
@@ -4,13 +4,9 @@ dist: trusty
 sudo: required
 addons:
   firefox: latest
+  chrome: stable
   sauce_connect: true
-  apt:
-    sources:
-      - google-chrome
-    packages:
-      - google-chrome-stable
-script: xvfb-run wct --module-resolution=node --npm
+script: xvfb-run wct
 env:
   global:
     - secure: nRjnNSJqCw6HBd01k6+jMOEhCsyxuzjrIGeYJDfOBqBu9UXCQYEYUJBzC+n9DkAqWfm2/X5TOsMPLSo83WNO/IXEZPBAVn4+JCw9A9UA6giaLCQQD89ljecDVCeQ+2LUhNysn6C9Jy1YBJhVUK8DCDp64+nsyjtmfKjDGKK9Vi1WSAzH/MBTKTrt9OlNojVZeVijShuFPuWn43NrM/YBHsETAaor1M/Z94NjVmSOiBdLO/y2o7XhYr5nbSDjOz38ok9A7azP2Chyi9gvuXjMLdP0vm3ha5LnDqC5hJDSIhd9TytezKLlM8yXGuaA+9W25qxWH4xeErygIEaJ8syBKAwEiyleQ0cD1nh++hlsEW9ER53YPnHf5RuudwrOBFVkJ03bKXlJmsjmSMzZNYfAVkXaCTw3Kg2j7lTCrLwgmOsld38OQW9VMhG+kgvjlTF/D+b6z6LscDsVcv3AOihZ3JYYajI/ULtrteFPzVQKY+Hjmh9d1V6ECg/ieeQJI3Ncubganj02P1+BnO6jFLC9mCHbZBftX1m8JnNRYPzPDPUxN8qgH130y/aexordoiYA1kCCtIZHM9jbh5+aTmCGVS8PzKZLLb7xmI2lZip09I5XedL5T9eCSCXUPJCPQKOhRWCdYda/LJ/yMXpaNE/RELpisKT3nKmtlSGb8nhpgGU=
