--- demo/customstyling.html
@@ -1,57 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <script src="../../@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
-        <script type="module" src="../src/mammooc-rating-widget.js"></script>
-        <script type="module" src="../../@polymer/polymer/polymer-element.js"></script>
-        <script type="module" src="../../@polymer/iron-demo-helpers/demo-snippet.js"></script>
-        <script type="module" src="../../@polymer/iron-demo-helpers/demo-pages-shared-styles.js"></script>
-        <!-- FIXME(polymer-modulizer):
-        These imperative modules that innerHTML your HTML are
-        a hacky way to be sure that any mixins in included style
-        modules are ready before any elements that reference them are
-        instantiated, otherwise the CSS @apply mixin polyfill won't be
-        able to expand the underlying CSS custom properties.
-        See: https://github.com/Polymer/polymer-modulizer/issues/154
-        -->
-        <script type="module">
-            const $_documentContainer = document.createElement('template');
-
-            $_documentContainer.innerHTML = `<custom-style>
-                        <style include="demo-pages-shared-styles">
-                            .vertical-section-container {
-                                max-width: 550px;
-                            }
-                        </style>
-                    </custom-style>`;
-
-            document.body.appendChild($_documentContainer.content);
-        </script>
-    </head>
-    <body>
-        <script type="module">
-            const $_documentContainer = document.createElement('template');
-
-            $_documentContainer.innerHTML = `<div class="vertical-section-container centered">
-                <h4>An element with custom styling</h4>
-                <demo-snippet class="centered-demo">
-                    <template>
-                        <style is="custom-style">
-                            mammooc-rating-widget.customStyling {
-                                --mammooc-rating-widget-primary-color: #5F8C1C;
-                                --mammooc-rating-widget-secondary-color: #999;
-                                --mammooc-rating-widget-background-color: #eee;
-                            }
-                        </style>
-                    <mammooc-rating-widget class="customStyling" provider="open_mammooc" course-id="1"></mammooc-rating-widget>
-                    </template>
-                </demo-snippet>
-                <a href="index.html">Back</a>
-            </div>`;
-
-            document.body.appendChild($_documentContainer.content);
-        </script>
-    </body>
-
-
-</html>
