+++ demo/mammooc-course-evaluation.html
@@ -1,0 +1,60 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="utf-8">
+    <meta content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes" name="viewport">
+
+    <title>mammooc-cumulative-rating demo</title>
+
+    <script src="../../@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+
+    <script src="../../@polymer/polymer/polymer-element.js" type="module"></script>
+    <script src="../../@polymer/iron-demo-helpers/demo-pages-shared-styles.js" type="module"></script>
+    <script src="../../@polymer/iron-demo-helpers/demo-snippet.js" type="module"></script>
+    <script type="module" src="../src/elements/mammooc-course-evaluation.js"></script>
+
+    <!-- Bug fix by polymer-modulizer:
+        These imperative modules that innerHTML your HTML are
+        a hacky way to be sure that any mixins in included style
+        modules are ready before any elements that reference them are
+        instantiated, otherwise the CSS @apply mixin polyfill won't be
+        able to expand the underlying CSS custom properties.
+        See: https://github.com/Polymer/polymer-modulizer/issues/154
+    -->
+    <script type="module">
+        import { html } from '@polymer/polymer/lib/utils/html-tag.js';
+
+        const template = html`<custom-style>
+            <style include="demo-pages-shared-styles">
+                .vertical-section-container {
+                  max-width: 500px;
+                }
+            </style>
+        </custom-style>`;
+
+        document.body.appendChild(template.content);
+    </script>
+</head>
+
+<body>
+<div class="vertical-section-container centered">
+    <h4>A course evaluation</h4>
+    <demo-snippet class="centered-demo">
+        <template>
+            <mammooc-course-evaluation provider="open_mammooc" course-id="1" per-page="3" rating-count="9">
+            </mammooc-course-evaluation>
+        </template>
+    </demo-snippet>
+
+    <h4>A course evaluation without empty list entries</h4>
+    <demo-snippet class="centered-demo">
+        <template>
+            <mammooc-course-evaluation provider="open_mammooc" course-id="1" per-page="3" rating-count="9" hide-empty-evaluations="true">
+            </mammooc-course-evaluation>
+        </template>
+    </demo-snippet>
+</div>
+</body>
+
+</html>
