+++ demo/mammooc-cumulative-rating.html
@@ -1,0 +1,52 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="utf-8">
+    <meta content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes" name="viewport">
+
+    <title>mammooc-cumulative-rating demo</title>
+
+    <script src="../../@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+
+    <script src="../../@polymer/polymer/polymer-element.js" type="module"></script>
+    <script src="../../@polymer/iron-demo-helpers/demo-pages-shared-styles.js" type="module"></script>
+    <script src="../../@polymer/iron-demo-helpers/demo-snippet.js" type="module"></script>
+    <script type="module" src="../src/elements/mammooc-cumulative-rating.js"></script>
+
+    <!-- Bug fix by polymer-modulizer:
+        These imperative modules that innerHTML your HTML are
+        a hacky way to be sure that any mixins in included style
+        modules are ready before any elements that reference them are
+        instantiated, otherwise the CSS @apply mixin polyfill won't be
+        able to expand the underlying CSS custom properties.
+        See: https://github.com/Polymer/polymer-modulizer/issues/154
+    -->
+    <script type="module">
+        import { html } from '@polymer/polymer/lib/utils/html-tag.js';
+
+        const template = html`<custom-style>
+            <style include="demo-pages-shared-styles">
+                .vertical-section-container {
+                  max-width: 500px;
+                }
+            </style>
+        </custom-style>`;
+
+        document.body.appendChild(template.content);
+    </script>
+</head>
+
+<body>
+<div class="vertical-section-container centered">
+    <h4>A cumulative rating</h4>
+    <demo-snippet class="centered-demo">
+        <template>
+            <mammooc-cumulative-rating cumulative-rating="3.6" rating-count="1337">
+            </mammooc-cumulative-rating>
+        </template>
+    </demo-snippet>
+</div>
+</body>
+
+</html>
