--- mammooc-rating-widget.js
@@ -1,1 +1,0 @@
-import './src/mammooc-rating-widget.js';
