+++ package.json
@@ -7,7 +7,22 @@
   "homepage": "https://github.com/mammooc/mammooc-rating-widget#readme",
   "name": "mammooc-rating-widget",
   "version": "3.0.1",
-  "flat": true,
+  "main": "src/all-imports.js",
+  "bugs": {
+    "url": "https://github.com/mammooc/mammooc-rating-widget/issues"
+  },
+  "license": "MIT",
+  "scripts": {
+    "test": "node_modules/web-component-tester/bin/wct test/index.html",
+    "test-firefox": "node_modules/web-component-tester/bin/wct test/index.html -l firefox",
+    "test-safari": "node_modules/web-component-tester/bin/wct test/index.html -l safari",
+    "test-chrome": "node_modules/web-component-tester/bin/wct test/index.html -l chrome",
+    "testdebugging": "node_modules/web-component-tester/bin/wct -p test/index.html",
+    "lint": "eslint src --fix && htmlhint --config .htmlhintrc demo/**/*.html index.html && polymer lint",
+    "serve": "polymer serve",
+    "bundle": "gulp bundle",
+    "analyze": "polymer analyze > analysis.json"
+  },
   "contributors": [
     {
       "name": "Paul Arndt"
@@ -25,74 +40,64 @@
       "name": "Sebastian Serth"
     }
   ],
-  "main": "mammooc-rating-widget.js",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/mammooc/mammooc-rating-widget/issues"
-  },
   "directories": {
     "test": "test"
   },
   "dependencies": {
     "@polymer/app-localize-behavior": "^3.0.1",
     "@polymer/iron-ajax": "^3.0.1",
-    "@polymer/iron-demo-helpers": "^3.0.2",
     "@polymer/iron-form": "^3.0.1",
     "@polymer/iron-form-element-behavior": "^3.0.1",
     "@polymer/iron-jsonp-library": "^3.0.1",
-    "@polymer/iron-list": "^3.0.1",
+    "@polymer/iron-list": "^3.1.0",
     "@polymer/iron-scroll-threshold": "^3.0.1",
     "@polymer/paper-button": "^3.0.1",
     "@polymer/paper-card": "^3.0.1",
     "@polymer/paper-checkbox": "^3.1.0",
-    "@polymer/paper-input": "^3.0.1",
+    "@polymer/paper-input": "^3.2.1",
     "@polymer/paper-radio-button": "^3.0.1",
     "@polymer/paper-radio-group": "^3.0.1",
     "@polymer/paper-styles": "^3.0.1",
     "@polymer/paper-toast": "^3.0.1",
-    "@polymer/polymer": "^3.1.0",
+    "@polymer/polymer": "^3.3.1",
     "flexible-rating": "^3.0.3",
-    "intl-messageformat": "^2.2.0"
+    "intl-messageformat": "^8.3.7"
   },
   "devDependencies": {
-    "@polymer/iron-component-page": "^3.0.0-pre.21",
-    "@webcomponents/webcomponentsjs": "^2.2.6",
-    "babel-core": "^6.26.3",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "eslint": "^5.3.0",
-    "eslint-plugin-html": "^4.0.5",
-    "gulp": "^4.0.0",
+    "@babel/core": "^7.9.0",
+    "@babel/polyfill": "^7.8.7",
+    "@babel/preset-env": "^7.9.5",
+    "@polymer/iron-component-page": "^4.0.1",
+    "@polymer/iron-demo-helpers": "^3.1.0",
+    "@polymer/iron-test-helpers": "^3.0.1",
+    "@polymer/test-fixture": "^4.0.2",
+    "@webcomponents/webcomponentsjs": "^2.4.3",
+    "chai": "^4.2.0",
+    "core-js": "^2.6.11",
+    "eslint": "^6.8.0",
+    "eslint-plugin-html": "^6.0.2",
+    "gulp": "^4.0.2",
+    "gulp-babel": "^8.0.0",
     "gulp-clean": "^0.4.0",
-    "gulp-cli": "^2.0.1",
+    "gulp-cli": "^2.2.0",
     "gulp-concat": "^2.6.1",
+    "gulp-if": "^3.0.0",
     "gulp-insert": "^0.5.0",
-    "gulp-rollup": "^2.16.2",
-    "gulp-uglify-es": "^1.0.4",
-    "htmlhint": "^0.9.13",
-    "polymer-bundler": "^4.0.2",
-    "polymer-cli": "^1.7.7",
-    "rollup-plugin-node-resolve": "^4.0.0",
-    "url-polyfill": "^1.1.3",
-    "wct-browser-legacy": "^1.0.2",
+    "gulp-rename": "^2.0.0",
+    "gulp-rollup": "^2.17.0",
+    "gulp-tap": "^2.0.0",
+    "gulp-terser-js": "^5.1.2",
+    "htmlhint": "^0.11.0",
+    "into-stream": "^5.1.1",
+    "merge-stream": "^2.0.0",
+    "minify-html-literals": "^1.3.0",
+    "mocha": "^7.1.1",
+    "polymer-build": "^3.1.4",
+    "polymer-cli": "^1.9.11",
+    "rollup-plugin-commonjs": "^9.3.4",
+    "rollup-plugin-node-resolve": "^4.2.4",
+    "vinyl-buffer": "^1.0.1",
+    "wct-mocha": "^1.0.1",
     "web-component-tester": "^6.9.2"
-  },
-  "scripts": {
-    "test": "node_modules/web-component-tester/bin/wct --module-resolution=node --npm test/index.html",
-    "test-firefox": "node_modules/web-component-tester/bin/wct --module-resolution=node --npm test/index.html -l firefox",
-    "test-safari": "node_modules/web-component-tester/bin/wct --module-resolution=node --npm test/index.html -l safari",
-    "test-chrome": "node_modules/web-component-tester/bin/wct --module-resolution=node --npm test/index.html -l chrome",
-    "testdebugging": "node_modules/web-component-tester/bin/wct --module-resolution=node --npm -p test/index.html",
-    "lint": "eslint src --fix && htmlhint --config .htmlhintrc src/**/*.html index.html && polymer lint",
-    "serve": "polymer serve --module-resolution=node --npm",
-    "bundle": "gulp bundle",
-    "analyze": "polymer analyze > analysis.json",
-    "deploy": "./utils/deploy.sh"
-  },
-  "resolutions": {
-    "inherits": "2.0.3",
-    "samsam": "1.1.3",
-    "supports-color": "3.1.2",
-    "type-detect": "1.0.0"
   }
 }
