+++ polymer.json
@@ -1,35 +1,16 @@
 {
-  "entrypoint": "mammooc-rating-widget.js",
+  "entrypoint": "src/all-imports.js",
   "sources": [
-    "src/**/*"
+    "src/**/*.js"
+  ],
+  "extraDependencies": [
+    "node_modules/@webcomponents/webcomponentsjs/*.js"
   ],
   "lint": {
     "rules": [
       "polymer-3"
-    ],
-    "ignoreWarnings": [
-      "unknown-superclass",
-      "set-unknown-attribute",
-      "databinding-calls-must-be-functions",
-      "databind-with-unknown-property"
     ]
   },
-  "builds": [
-    {
-      "name": "es6",
-      "bundle": true,
-      "js": {"minify": true},
-      "css": {"minify": true},
-      "html": {"minify": true}
-    },
-    {
-      "name": "es5",
-      "bundle": true,
-      "js": {"minify": true, "compile": true},
-      "css": {"minify": true},
-      "html": {"minify": true}
-    }
-  ],
   "moduleResolution": "node",
   "npm": true
 }
