+++ src/all-imports.js
@@ -1,0 +1,5 @@
+/* Import all components: all-imports.js */
+import './elements/mammooc-course-evaluation.js';
+import './elements/mammooc-course-evaluation-form.js';
+import './elements/mammooc-cumulative-rating.js';
+import './elements/mammooc-rating-widget.js';
