+++ src/elements/mammooc-cumulative-rating.js
@@ -1,30 +1,24 @@
-
-import { PolymerElement } from '@polymer/polymer/polymer-element.js';
-
-import '@polymer/polymer/lib/elements/custom-style.js';
-import '@webcomponents/shadycss/entrypoints/apply-shim.js';
-import 'flexible-rating/flexible-rating.js';
 import '@polymer/paper-styles/typography.js';
-import { MammoocLocalizeBehavior } from './mammooc-localize-behavior.js';
-import { html } from '@polymer/polymer/lib/utils/html-tag.js';
-import { mixinBehaviors } from '@polymer/polymer/lib/legacy/class.js';
+import '@polymer/polymer/lib/elements/custom-style.js';
+import 'flexible-rating/flexible-rating.js';
+import { MammoocLocalizeMixin } from '../mixins/mammooc-localize-mixin.js';
+import { PolymerElement, html } from '@polymer/polymer';
 
 /**
-An element providing an overall rating.
-
-Example:
-
-    <mammooc-cumulative-rating></mammooc-cumulative-rating>
-
-@element mammooc-cumulative-rating
-*/
-class MammoocCumulativeRating extends mixinBehaviors([
-    MammoocLocalizeBehavior
-], PolymerElement) {
+ * ## `mammooc-cumulative-rating`
+ *
+ * An element providing an overall rating.
+ *
+ * @polymer
+ * @customElement
+ * @appliesMixin MammoocLocalizeMixin
+ * @demo demo/mammooc-cumulative-rating.html
+ */
+class MammoocCumulativeRating extends MammoocLocalizeMixin(PolymerElement) {
     static get template() {
         return html`
-        <custom-style>
-            <style>
+        <custom-style> <!-- custom-style tag is required as Polyfill for otherwise unsupported browsers, such as IE 11 -->
+            <style include="paper-material-styles">
                 :host {
                     --iron-icon-height: 50px;
                     --iron-icon-width: 50px;
@@ -36,7 +30,6 @@ class MammoocCumulativeRating extends mixinBehaviors([
             </style>
         </custom-style>
 
-        <!-- local DOM goes here -->
         <div class="heading">
             {{localize("ratings")}}
         <flexible-rating value="{{cumulativeRating}}" disabled=""></flexible-rating> ({{ratingCount}})
@@ -50,18 +43,11 @@ class MammoocCumulativeRating extends mixinBehaviors([
 
     static get properties() {
         return {
-          /**
-           *   The average rating for the course
-           *   @type {number}
-           */
+            /** The average rating for the course */
             cumulativeRating: {
                 type: Number
             },
-          
-          /**
-           *   The number of ratings for the course
-           *   @type {number}
-           */
+            /** The number of ratings for the course */
             ratingCount: {
                 type: Number
             }
@@ -70,10 +56,6 @@ class MammoocCumulativeRating extends mixinBehaviors([
 
     static get observers() {
         return [];
-    }
-
-    ready() {
-        super.ready();
     }
 }
