+++ test/index.html
@@ -1,17 +1,24 @@
 <!doctype html>
 <html>
-  <head>
+<head>
     <meta charset="utf-8">
+    <meta content="width=device-width, minimum-scale=1, initial-scale=1, user-scalable=yes" name="viewport">
+
+    <title>mammooc-rating-widget test</title>
+
     <script src="../../@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
-    <script src="../../wct-browser-legacy/browser.js"></script>
-  </head>
-  <body>
-    <script>
-      WCT.loadSuites([
+    <script src="../node_modules/@polymer/test-fixture/test-fixture.js"></script>
+    <script src="../node_modules/mocha/mocha.js"></script>
+    <script src="../node_modules/chai/chai.js"></script>
+    <script src="../node_modules/wct-mocha/wct-mocha.js"></script>
+</head>
+<body>
+<script>
+    WCT.loadSuites([
         'properties.html',
         'form.html',
         'structure.html'
-      ]);
-    </script>
-  </body>
+    ]);
+</script>
+</body>
 </html>
